+++ babel.config.js
@@ -1,0 +1,17 @@
+module.exports = {
+    "presets": [
+        [
+            "@babel/preset-env",
+            {
+                "targets": {
+                    "node": true
+                }
+            }
+        ]
+    ],
+    "ignore": [
+        "node_modules/",
+        "**/dist/",
+        "**/*.dist.js"
+    ]
+};