+++ projects/Simple-React/src/components/Index/Index.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Index = () => (
+  <h3>Index page</h3>
+)
+
+export default Index