+++ projects/Simple-React/src/components/Page/index.js
@@ -1,0 +1,3 @@
+import Page from './Page.jsx'
+
+export default Page